Made some new files

+++ Timer.c
@@ -3,7 +3,7 @@
 #include <stdint.h>
 /*Methods for time handeling in the game*/
 
-void quickTimer(uint8_t timeout) {
+void quickTimer(int timeout) {
 
     while (timeout) {
         timeout--;
